refactor(ast): keep the macro engine on root, presets on @kubb/ast/macros
The macro engine (defineMacro, composeMacros, applyMacros, and the Macro
type) lives on the @kubb/ast root barrel again. The @kubb/ast/macros
subpath now holds only the built-in macro presets (macroDiscriminatorEnum,
macroSimplifyUnion, macroEnumName). Core imports the engine from the root,
and the OAS parser uses ast.applyMacros with presets from @kubb/ast/macros.
