+++ packages/ast/src/index.ts
@@ -4,6 +4,8 @@ export { defineSchemaDialect } from './dialect.ts'
 // Node constructors, also published as the `@kubb/ast/factory` subpath, mirroring `ts.factory.createX`.
 export * as factory from './factory.ts'
 export { isHttpOperationNode, narrowSchema } from './guards.ts'
+export { applyMacros, composeMacros, defineMacro } from './macro.ts'
+export type { Macro } from './macro.ts'
 export { defineNode } from './node.ts'
 export type { NodeDef } from './node.ts'
 export { syncOptionality } from './node.ts'
