+++ packages/ast/src/macro.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it } from 'vitest'
-import { createInput } from '../nodes/input.ts'
-import { createOperation } from '../nodes/operation.ts'
-import { createProperty } from '../nodes/property.ts'
-import { createSchema } from '../nodes/schema.ts'
-import { transform } from '../visitor.ts'
-import { applyMacros, composeMacros, defineMacro } from './defineMacro.ts'
+import { createInput } from './nodes/input.ts'
+import { createOperation } from './nodes/operation.ts'
+import { createProperty } from './nodes/property.ts'
+import { createSchema } from './nodes/schema.ts'
+import { applyMacros, composeMacros, defineMacro } from './macro.ts'
+import { transform } from './visitor.ts'
 
 describe('defineMacro', () => {
   it('returns the macro unchanged', () => {
