+++ packages/ast/src/macros/index.ts
@@ -1,3 +1,3 @@
-export { macroDiscriminatorEnum, macroEnumName, macroSimplifyUnion } from './builtins.ts'
-export { applyMacros, composeMacros, defineMacro } from './defineMacro.ts'
-export type { Macro } from './defineMacro.ts'
+// Built-in macro presets. The macro engine (`defineMacro`, `composeMacros`, `applyMacros`, and the
+// `Macro` type) lives on the `@kubb/ast` root barrel.
+export { macroDiscriminatorEnum, macroEnumName, macroSimplifyUnion } from './presets.ts'
