+++ packages/ast/src/macros/presets.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest'
+import { applyMacros } from '../macro.ts'
 import { createProperty } from '../nodes/property.ts'
 import { createSchema, type SchemaNode } from '../nodes/schema.ts'
-import { applyMacros } from './defineMacro.ts'
-import { macroDiscriminatorEnum, macroEnumName, macroSimplifyUnion } from './builtins.ts'
+import { macroDiscriminatorEnum, macroEnumName, macroSimplifyUnion } from './presets.ts'
 
 function applyShallow(node: SchemaNode, macro: Parameters<typeof applyMacros>[1][number]): SchemaNode {
   return applyMacros(node, [macro], { depth: 'shallow' })
